requirements.txt
pydantic
PyYAML
httpx
uvicorn
fastapi
loguru
pydantic-settings-yaml
aiofiles
